requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
jose==1.0.0
Naked==0.1.31
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.7.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
status==0.2.5
tomli==2.0.1
tornado==6.2
trio==0.21.0
typing-extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
